requirements.txt
PyPDF2
sentence-transformers
faiss-cpu
transformers
streamlit
